Enforce DCP ownership, index bounds, and stable plan identity
DCP planning validates runtime ownership and prepared capacity. Int64 index expansion protects large merged block IDs, compile admission keys selections without discarding factory inputs, and PCIe transport policies remain frozen after preparation.

+++ b12x/attention/dsa_indexer/mxfp4.py
@@ -840,7 +840,7 @@ class _SortPositions:
             if cutlass.const_expr(not self.expand_blocks):
                 valid = valid and value > Float32(-float("inf"))
             if cutlass.const_expr(self.expand_blocks):
-                valid = valid and idx * Int32(8) < visible
+                valid = valid and Int64(idx) * Int64(8) < Int64(visible)
             else:
                 valid = valid and idx < visible
             if not valid:
